+++ CHANGES.md
@@ -1,5 +1,6 @@
 # UNRELEASED
 - [NEW] Allow a custom HTTP agent to be specified.
+- [NEW] Trim the database URL if it includes the `/_changes` endpoint.
 - [FIXED] Retry `/_db_updates` request using the last recorded sequence value as
   the `since` query parameter.